requirements.txt
tensorboard
numpy
torchsummary
scikit-image